requirements.txt
autopep8==1.5.7
pycodestyle==2.7.0
python-dotenv==0.19.0
PyYAML==5.4.1
toml==0.10.2
